requirements.txt
# requirements.txt
# ====================================================================
# Crypto Bot v52.4 - 本番環境最小依存関係（2025年11月更新）
# Phase 52.4-B完了: コード品質改善・6戦略55特徴量システム・Phase参照統一67%削減
# ====================================================================

# --- 数値計算・データ処理基盤 ----------------------------------------
numpy>=2.0.0            # 数値計算基盤・Python3.13対応・プリビルドwheel利用・高速インストール
pandas>=2.0.0           # データフレーム処理・Python3.13対応・numpy2.x互換・時系列データ分析最適化
scikit-learn==1.7.1     # 機械学習ライブラリ・CI/CD環境統一・モデル互換性確保
joblib==1.3.2           # 並列処理・モデルシリアライゼーション

# --- HTTP通信・API接続 ------------------------------------------------
requests==2.31.0        # HTTP クライアント・外部API通信
ccxt>=2.10.0            # 暗号資産取引所API・Bitbank接続・版互換性確保
yfinance==0.2.32        # Yahoo Finance API・Phase 50.3外部API統合・マクロ経済指標取得
aiohttp>=3.9.0          # 非同期HTTPクライアント・Phase 50.3外部API統合・Alternative.me API接続

# --- Cloud Storage・永続化 -------------------------------------------
google-cloud-storage==2.10.0  # GCP Cloud Storage・ドローダウン状態永続化・GCP統合対応

# --- 設定・環境管理 ---------------------------------------------------
python-dotenv==1.0.0    # .env ファイル読み込み・環境変数管理
PyYAML>=6.0.1           # YAML設定ファイル処理・Python3.13対応・production.yml読み込み

# --- リトライ・エラーハンドリング --------------------------------------
tenacity==8.2.0         # リトライ機構・API呼び出し安定化

# --- Web API フレームワーク ------------------------------------------
fastapi>=0.100.0        # FastAPI フレームワーク・ヘルスチェックAPI・互換性確保
uvicorn>=0.20.0         # ASGI サーバー・FastAPI実行環境・互換性確保

# --- テクニカル指標計算 -----------------------------------------------
pandas-ta>=0.3.14b0     # テクニカル指標計算・TA-Lib代替・本番必須

# --- 機械学習モデル（本番環境必須・互換性保証版）---------------------
lightgbm>=4.5.0,<5.0.0  # LightGBM・アンサンブル学習・Python3.13対応・後方互換保証
xgboost>=3.0.0,<4.0.0   # XGBoost・アンサンブル学習・Python3.13対応・既存モデル互換性保証

# --- Phase 39.4-39.5: ML信頼度向上・最適化 --------------------------------
imbalanced-learn>=0.11.0  # SMOTEオーバーサンプリング・クラス不均衡対策・Phase 39.4
optuna>=3.3.0              # ハイパーパラメータ最適化・ベイズ最適化・Phase 39.5

# --- Phase 49: バックテスト可視化 -------------------------------------------
matplotlib>=3.8.0          # グラフ生成・エクイティカーブ・損益分布可視化・Phase 49.4
Pillow>=10.0.0             # 画像処理・matplotlib依存関係・Phase 49.4

# --- テスト・開発環境（開発時のみ）----------------------------------
pytest>=7.0.0          # テストフレームワーク・653テスト100%成功達成
pytest-cov>=4.0.0      # カバレッジ測定・58.62%カバレッジ・.cache/coverage統合
pytest-asyncio>=0.21.0 # 非同期テスト対応
black>=23.0.0           # コードフォーマッター・統一設定
isort>=5.12.0           # インポート整理・black連携
flake8>=6.0.0           # リンター・コード品質管理・実用的設定
mypy>=1.5.0             # 静的型チェッカー・Python3.13対応・実用的バランス

# --- Cloud Run・本番環境（GCP最適化）-------------------------------
gunicorn>=21.0.0        # WSGI サーバー・本番環境
uvloop>=0.17.0          # Linux高速非同期処理

# --- セキュリティ・暗号化 -------------------------------------------
cryptography>=41.0.0   # 暗号化・セキュリティ機能